+++ converter-dsl/src/main/java/linc/com/heifconverter/dsl/extension/create/CreateDslKts.kt
@@ -1,10 +1,13 @@
-package linc.com.heifconverter.dsl
+package linc.com.heifconverter.dsl.extension.create
 
 import android.content.Context
 import android.graphics.Bitmap
 import androidx.annotation.DrawableRes
 import linc.com.heifconverter.HeifConverter
 import linc.com.heifconverter.HeifConverter.Input
+import linc.com.heifconverter.dsl.HeifConverterDsl
+import linc.com.heifconverter.dsl.HeifConverterInstance
+import linc.com.heifconverter.dsl.InternalHeifConverterDsl
 import java.io.File
 import java.io.InputStream
